Added the abiliuty to search for Owners
